Split the protocol buffers up and tightened things up a bit.
The Packet is now just a routing envelope and we have removed
everything that isn't related to routing from it.  Serial numbers
and acknowledgements are now moved into the payload families and
we have decided to rename "id" to "serial" to make the intent
somewhat clearer.

This also maps better to a server and multiplexing structure. Especially
when we start work on things like file transfers, which we expect
to be the most complicated use-case initially, since it will have to
work across a wide range of transports over different kinds of
connectivity.